requirements.txt
Flask==2.0.1
qrcode==7.3.1
Pillow==9.3.0
python-dotenv==0.19.0
